dh_sameversiondeps: a helper untility which expands appropriately formatted substvar to the dependency which versioning is based on the versioning of other dependency (which most likely were generated by dpkg-shlibdeps).
E.g. in kdepimlibs5-dev case, it needs a pretty tightly (e.g. the shlibs version of kde4libs it was built against) versioned dependency on kdelibs5-dev. Therefore, the idea is to automatically add such kdelibs5-dev depencency that it is versioned in the same way as kdelibs5 dep is versioned for the package kdepimlibs5. Hence the  substvar ${sameVersionDep:kdelibs5-dev:kdepimlibs5-Depends-kdelibs5} on kdepimlibs5-dev  tells dh_sameversiondeps to:

* Get Depends field of the kdepimlibs5 package from debian/control, expand other substvars in that field based on the contents of debian/$package.substvars and debian/substvars.
* Find all kdelibs5 occurenciers (versioning included).
* s/kdelibs5/kdelibs5-dev/ and set the value of sameVersionDep:kdelibs5-dev:kdepimlibs5-Depends-kdelibs5 substvar to the resulting string dependency string. substvar is written to debian/$package.substvars by default.